# docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MemWire API",
    "description": "Persistent AI memory infrastructure \u2014 store, recall, and search memories for your AI agents.",
    "version": "1.0.0",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/v1/memory": {
      "post": {
        "summary": "Store memory",
        "description": "Store one or more conversation messages into the user's memory graph. Each message is embedded, classified, and indexed \u2014 ready for recall and search.",
        "operationId": "store_memory",
        "tags": [
          "Memory"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreRequest"
              },
              "example": {
                "user_id": "alice",
                "app_id": "my_app",
                "workspace_id": "team_1",
                "messages": [
                  {
                    "role": "user",
                    "content": "I prefer dark mode and short answers."
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Memories stored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                },
                "example": {
                  "stored": 1,
                  "memory_ids": [
                    "mem_3f7a1c2d9e4b"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/memory/recall": {
      "post": {
        "summary": "Recall memory context",
        "description": "Traverse the memory graph via BFS to find the most relevant context for a natural-language query. Returns a formatted string ready to inject into your LLM prompt.",
        "operationId": "recall_memory",
        "tags": [
          "Memory"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecallRequest"
              },
              "example": {
                "user_id": "alice",
                "app_id": "my_app",
                "workspace_id": "team_1",
                "query": "How should I format my answers?",
                "top_k": 5
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recalled context and matching paths",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallResponse"
                },
                "example": {
                  "context": "alice prefers dark mode and short answers.",
                  "paths": 2,
                  "knowledge": []
                }
              }
            }
          }
        }
      }
    },
    "/v1/memory/search": {
      "post": {
        "summary": "Search memories",
        "description": "Semantic similarity search over stored memories using hybrid dense + sparse retrieval. Optionally filter by category.",
        "operationId": "search_memory",
        "tags": [
          "Memory"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              },
              "example": {
                "user_id": "alice",
                "query": "dark mode",
                "top_k": 10
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of matching memories with similarity scores",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                },
                "example": {
                  "results": [
                    {
                      "memory_id": "mem_3f7a1c2d9e4b",
                      "content": "I prefer dark mode and short answers.",
                      "category": "preference",
                      "score": 0.94
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/knowledge": {
      "post": {
        "summary": "Add knowledge base",
        "description": "Ingest a named knowledge base of document chunks. Chunks are embedded and indexed for semantic search and recall.",
        "operationId": "add_knowledge",
        "tags": ["Knowledge"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddKnowledgeRequest" },
              "example": {
                "user_id": "alice",
                "name": "Company Handbook",
                "chunks": [
                  { "content": "All employees must complete onboarding within 30 days." },
                  { "content": "Remote work is allowed up to 3 days per week." }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Knowledge base created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AddKnowledgeResponse" },
                "example": { "kb_id": "kb_3f7a1c2d9e4b", "chunks_stored": 2 }
              }
            }
          }
        }
      }
    },
    "/v1/knowledge/search": {
      "post": {
        "summary": "Search knowledge",
        "description": "Search knowledge base chunks by semantic similarity using hybrid dense+sparse retrieval.",
        "operationId": "search_knowledge",
        "tags": ["Knowledge"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/KnowledgeSearchRequest" },
              "example": {
                "user_id": "alice",
                "query": "remote work policy",
                "top_k": 5
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matching knowledge chunks",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/KnowledgeSearchResponse" },
                "example": {
                  "chunks": [
                    {
                      "chunk_id": "kb_3f7a1c2d9e4b_chunk_1",
                      "kb_id": "kb_3f7a1c2d9e4b",
                      "content": "Remote work is allowed up to 3 days per week.",
                      "score": 0.92
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/knowledge/{kb_id}": {
      "delete": {
        "summary": "Delete knowledge base",
        "description": "Delete a knowledge base and all its chunks by kb_id.",
        "operationId": "delete_knowledge",
        "tags": ["Knowledge"],
        "parameters": [
          {
            "name": "kb_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "The knowledge base ID returned when the KB was created."
          }
        ],
        "responses": {
          "200": {
            "description": "Knowledge base deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteKnowledgeResponse" },
                "example": { "deleted": true }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Returns the server status. Use this to verify the API is up before sending requests.",
        "operationId": "health_check",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "Server is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Message": {
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant",
              "system"
            ],
            "description": "The role of the message author."
          },
          "content": {
            "type": "string",
            "description": "The text content of the message."
          }
        }
      },
      "StoreRequest": {
        "type": "object",
        "required": [
          "user_id",
          "messages"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "description": "User identifier. Required for all operations."
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "One or more conversation messages to store."
          },
          "app_id": {
            "type": "string",
            "description": "Optional application namespace.",
            "nullable": true
          },
          "workspace_id": {
            "type": "string",
            "description": "Optional workspace or team namespace.",
            "nullable": true
          },
          "agent_id": {
            "type": "string",
            "description": "Optional agent namespace.",
            "nullable": true
          }
        }
      },
      "StoreResponse": {
        "type": "object",
        "properties": {
          "stored": {
            "type": "integer",
            "description": "Number of messages stored."
          },
          "memory_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "IDs of the created memory records."
          }
        }
      },
      "RecallRequest": {
        "type": "object",
        "required": [
          "user_id",
          "query"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "description": "User to recall memory for."
          },
          "query": {
            "type": "string",
            "description": "Natural-language query to find relevant memories."
          },
          "app_id": {
            "type": "string",
            "nullable": true
          },
          "workspace_id": {
            "type": "string",
            "nullable": true
          },
          "agent_id": {
            "type": "string",
            "nullable": true
          },
          "top_k": {
            "type": "integer",
            "default": 5,
            "description": "Maximum number of paths to return."
          }
        }
      },
      "RecallResponse": {
        "type": "object",
        "properties": {
          "context": {
            "type": "string",
            "description": "Formatted context string ready to inject into an LLM prompt."
          },
          "paths": {
            "type": "integer",
            "description": "Number of supporting memory paths found."
          },
          "knowledge": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Matching knowledge base chunks, if any."
          }
        }
      },
      "SearchRequest": {
        "type": "object",
        "required": [
          "user_id",
          "query"
        ],
        "properties": {
          "user_id": {
            "type": "string"
          },
          "query": {
            "type": "string",
            "description": "Semantic search query."
          },
          "app_id": {
            "type": "string",
            "nullable": true
          },
          "workspace_id": {
            "type": "string",
            "nullable": true
          },
          "agent_id": {
            "type": "string",
            "nullable": true
          },
          "top_k": {
            "type": "integer",
            "default": 10,
            "description": "Number of results to return."
          },
          "category": {
            "type": "string",
            "nullable": true,
            "description": "Filter by memory category: preference, fact, instruction, event, or entity.",
            "enum": [
              "preference",
              "fact",
              "instruction",
              "event",
              "entity"
            ]
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "memory_id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "number",
            "format": "float",
            "description": "Similarity score between 0 and 1."
          }
        }
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            }
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        }
      },
      "KnowledgeChunkInput": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": {
            "type": "string",
            "description": "The text content of this chunk."
          },
          "metadata": {
            "type": "object",
            "description": "Optional metadata (e.g. source, page number).",
            "additionalProperties": true
          }
        }
      },
      "AddKnowledgeRequest": {
        "type": "object",
        "required": ["user_id", "name", "chunks"],
        "properties": {
          "user_id": { "type": "string", "description": "User identifier." },
          "name": { "type": "string", "description": "Human-readable name for the knowledge base." },
          "chunks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/KnowledgeChunkInput" },
            "description": "List of document chunks to ingest."
          },
          "agent_id": { "type": "string", "nullable": true },
          "app_id": { "type": "string", "nullable": true },
          "workspace_id": { "type": "string", "nullable": true }
        }
      },
      "AddKnowledgeResponse": {
        "type": "object",
        "properties": {
          "kb_id": { "type": "string", "description": "The ID of the created knowledge base." },
          "chunks_stored": { "type": "integer", "description": "Number of chunks ingested." }
        }
      },
      "KnowledgeSearchRequest": {
        "type": "object",
        "required": ["user_id", "query"],
        "properties": {
          "user_id": { "type": "string" },
          "query": { "type": "string", "description": "Natural-language search query." },
          "agent_id": { "type": "string", "nullable": true },
          "app_id": { "type": "string", "nullable": true },
          "workspace_id": { "type": "string", "nullable": true },
          "top_k": { "type": "integer", "default": 5, "description": "Number of chunks to return." }
        }
      },
      "KnowledgeChunkResult": {
        "type": "object",
        "properties": {
          "chunk_id": { "type": "string" },
          "kb_id": { "type": "string" },
          "content": { "type": "string" },
          "score": { "type": "number", "format": "float", "description": "Similarity score between 0 and 1." }
        }
      },
      "KnowledgeSearchResponse": {
        "type": "object",
        "properties": {
          "chunks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/KnowledgeChunkResult" }
          }
        }
      },
      "DeleteKnowledgeResponse": {
        "type": "object",
        "properties": {
          "deleted": { "type": "boolean", "description": "True if the knowledge base was deleted." }
        }
      }
    }
  }
}